Changed Auth flow

+++ api/views.py
@@ -1,3 +1,4 @@
+from django.shortcuts import get_object_or_404
 from rest_framework import generics, permissions, status
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
@@ -7,7 +8,7 @@ from django.contrib.auth import get_user_model
 from .serializers import CustomUserSerializer, UserSerializer, EventSerializer
 from .models import Event
 from .tasks import  send_event_notification
-
+from django.contrib.auth import authenticate
 class UserRegistration(APIView):
     permission_classes = [permissions.AllowAny]
 
@@ -36,8 +37,22 @@ class UserLogin(APIView):
             'access': str(refresh.access_token),
         }
         return Response(response_data, status=status.HTTP_200_OK)
- # api/views.py
+class RefreshTokenView(APIView):
+    permission_classes = [permissions.AllowAny]
 
+    def post(self, request):
+        refresh_token = request.data.get('refresh_token')
+
+        if refresh_token is None:
+            return Response({'error': 'Refresh token is required.'}, status=status.HTTP_400_BAD_REQUEST)
+
+        try:
+            refresh = RefreshToken(refresh_token)
+            access_token = str(refresh.access_token)
+            return Response({'access': access_token}, status=status.HTTP_200_OK)
+        except Exception as e:
+            return Response({'error': str(e)}, status=status.HTTP_401_UNAUTHORIZED)
+        
 @api_view(['GET', 'POST', 'DELETE', 'PUT'])
 @permission_classes([permissions.IsAuthenticated])
 def event_list(request, pk=None):
@@ -57,30 +72,40 @@ def event_list(request, pk=None):
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     elif request.method == 'DELETE':
-        event = Event.objects.get(pk=pk)
-        event_details = {
-            "id": event.id,
-            "title": event.title,
-            "description": event.description,
-            "date": event.date,
-            "time": event.time,
-            "location": event.location
-        }
+        event = get_object_or_404(Event, pk=pk)
         
-        event.delete()
+        # Check if the requesting user is the creator of the event
+        if request.user == event.creator:
+            event_details = {
+                "id": event.id,
+                "title": event.title,
+                "description": event.description,
+                "date": event.date,
+                "time": event.time,
+                "location": event.location
+            }
 
-        # Asynchronously send event notification email using Celery
-        send_event_notification.apply_async(args=[event_details, 'DELETE'])
+            event.delete()
 
-        return Response(status=status.HTTP_204_NO_CONTENT)
+            # Asynchronously send event notification email using Celery
+            send_event_notification.apply_async(args=[event_details, 'DELETE'])
+
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response({'error': 'You do not have permission to delete this event.'}, status=status.HTTP_403_FORBIDDEN)
 
     elif request.method == 'PUT':
-        event = Event.objects.get(pk=pk)
-        serializer = EventSerializer(event, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
+        event = get_object_or_404(Event, pk=pk)
+        
+        # Check if the requesting user is the creator of the event
+        if request.user == event.creator:
+            serializer = EventSerializer(event, data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
 
-        # Asynchronously send event notification email using Celery
-        send_event_notification.apply_async(args=[pk, 'PUT'])
+            # Asynchronously send event notification email using Celery
+            send_event_notification.apply_async(args=[pk, 'PUT'])
 
-        return Response(serializer.data)
+            return Response(serializer.data)
+        else:
+            return Response({'error': 'You do not have permission to update this event.'}, status=status.HTTP_403_FORBIDDEN)
